switch to using `gonz-aurora.vim` instead of the ruby file
This is mostly because it's become clear that I should use treesitter
scopes for the highlighting and while this is possible in Neovim, it's
not supported by `palette`, the Ruby gem that I use to generate the vim
files.

There are things in `palette` that I want, so I'll probably reimplement
it and add support for more freeform scope names, etc.